+++ src/snowflake/cli/api/config.py
@@ -83,6 +83,14 @@ class ConnectionConfig:
     authenticator: Optional[str] = None
     private_key_file: Optional[str] = None
     token_file_path: Optional[str] = None
+    oauth_client_id: Optional[str] = None
+    oauth_client_secret: Optional[str] = None
+    oauth_authorization_url: Optional[str] = None
+    oauth_token_request_url: Optional[str] = None
+    oauth_redirect_uri: Optional[str] = None
+    oauth_scope: Optional[str] = None
+    oauth_security_features: Optional[str] = None
+    client_store_temporary_credential: Optional[bool] = None
 
     _other_settings: dict = field(default_factory=lambda: {})